Jfrog cli artifactory download - after corrections
